requirements.txt
transformers==3.1.0
pytorch-lightning==0.9.0
pytorch-nlp==0.5.0
sacrebleu==1.4.14
bert-score==0.3.6